finally

+++ English_System/src/main/java/com/wynne/ServiceImpl/DownloadServiceImpl.java
@@ -8,12 +8,15 @@ import org.springframework.stereotype.Service;
 import com.wynne.Dao.DownloadMapper;
 import com.wynne.Entity.Download;
 import com.wynne.Serivce.IDownloadService;
+import com.wynne.Utils.deleteFile;
 import com.wynne.Utils.fileRename;
 @Service
 public class DownloadServiceImpl implements IDownloadService{
 
 	@Autowired
 	private DownloadMapper downloadMapper;
+
+	private static final String path="G:/毕设/temp/";
 
 	public int insertrecord(Download record) {
 		int count=0;
@@ -27,41 +30,60 @@ public class DownloadServiceImpl implements IDownloadService{
 		return downloadMapper.selectAll();
 	}
 
+	public List<Download> findAll2(String filetype){
+		return downloadMapper.selectAllOption(filetype);
+	}
+
 	public int deletedownload(int fileid) {
-		return downloadMapper.deleteByPrimaryKey(fileid);
+		Download download=new Download();
+		int count=0;
+		int count1=0;
+		if(fileid>0){
+			download=downloadMapper.selectByPrimaryKey(fileid);
+			String filepath=download.getFilepath();
+			String realfilePath=path+filepath.substring(6, filepath.length());
+			System.out.println(realfilePath);
+			count=deleteFile.delete(realfilePath);		
+		}
+		if(count==1){
+			count1=downloadMapper.deleteByPrimaryKey(fileid);
+		}
+		return count1;
 	}
 
 	public Download findDownloadByPrimaryKey(int fileid) {
 		return downloadMapper.selectByPrimaryKey(fileid);
 	}
 
+
+
 	public int Update(Download download) {
+		return downloadMapper.updateByPrimaryKeySelective(download);
+	}
+
+	public List<Download> findAllOption2(String filetype, String filetypes) {
+		return downloadMapper.selectAllOption2(filetype, filetypes);
+	}
+
+	public int Update2(Download download) {
 		int count=0;
+		System.out.println(download.toString());
 		if(download!=null){
 			String filename=download.getFilename();
-			String filername=download.getFilername();
+			String filerealname=download.getFilerealname();
 			String rpath=download.getFilepath();
+			System.out.println(rpath);
 			int index=filename.lastIndexOf(".");
 			String str=filename.substring(index,filename.length());
-			if(filername.contains(".")){
-				int index2=filername.lastIndexOf(".");
-				String str2=filername.substring(index2, filername.length());
-				System.out.println(str.equals(str2));
-				if(!str.equals(str2)){
-					filername=filername+str;
-				}
-			}else{
-				filername=filername+str;
-			}
-			download.setFilername(filername);
-			download.setFilename(download.getFilename().substring(0, 14)+filername);
-			int index3=download.getFilepath().lastIndexOf("/")+14;
-			download.setFilepath(download.getFilepath().substring(0, index3)+filername);
-			if(fileRename.rename(rpath, download.getFilepath())){
+			download.setFilename(filerealname+str);
+			int index2=download.getFilepath().lastIndexOf("/");
+			download.setFilepath(download.getFilepath().substring(0, index2+1)+download.getFilename());
+			if(fileRename.rename(rpath, download.getFilename())){
 				count=downloadMapper.updateByPrimaryKeySelective(download);
 			}
 		}
 		return count;
 	}
-
 }
+
+
